test: apply review cleanup findings for compare and routing tests
- Dedupe MeteredFakeClient into tests/fake_llm.py shared by orchestrator
  and autofix tests
- Collapse redundant assertIsNotNone/assert narrowing pairs
- Reduce non-string-role regression to the single planned '[]' case
- Pin interrupted review call count to the exact deterministic value
- Derive autofix token expectation from METERED_TOTAL_TOKENS
