+++ tests/test_orchestrator.py
@@ -10,7 +10,7 @@ import unittest
 from pathlib import Path
 from unittest.mock import patch
 
-from tests.fake_llm import routing_handler
+from tests.fake_llm import MeteredFakeClient, routing_handler
 from tests.sample_data import write_sample_epub, write_sample_txt
 from trans_novel.agents.reviewer import ReviewOutputError
 from trans_novel.config import Config
@@ -96,35 +96,6 @@ def _config(state_dir: str):
             "paths": {"state_dir": state_dir},
         }
     )
-
-
-class MeteredFakeClient(FakeClient):
-    """Record small usage per offline call to verify review accounting isolation."""
-
-    def complete(
-        self,
-        messages,
-        *,
-        operation,
-        json_mode=False,
-        max_tokens=None,
-    ):
-        self.usage.record(
-            self.routes[operation].tier or "direct",
-            UsageSample(
-                prompt_tokens=5,
-                completion_tokens=3,
-                total_tokens=8,
-                cache_miss_tokens=5,
-            ),
-            operation,
-        )
-        return super().complete(
-            messages,
-            operation=operation,
-            json_mode=json_mode,
-            max_tokens=max_tokens,
-        )
 
 
 class TestOrchestrator(unittest.TestCase):
@@ -1793,7 +1764,7 @@ class TestReviewReporting(unittest.TestCase):
             review_dir = os.path.join(store.reviews_dir, sorted(os.listdir(store.reviews_dir))[-1])
             run1_review_calls = len(client1.calls) - translated_calls
             # The interrupted call still reached the transport and must be billed once.
-            self.assertGreaterEqual(run1_review_calls, 2)
+            self.assertEqual(run1_review_calls, 3)
             with open(os.path.join(review_dir, "usage.json"), encoding="utf-8") as file:
                 interrupted_usage = json.load(file)
             self.assertEqual(interrupted_usage["totals"]["calls"], run1_review_calls)
@@ -1817,7 +1788,6 @@ class TestReviewReporting(unittest.TestCase):
                 run1_review_calls + run2_review_calls,
             )
             book_usage = store.load_usage()
-            self.assertIsNotNone(book_usage)
             assert book_usage is not None
             self.assertEqual(
                 book_usage["totals"]["calls"],
